requirements.txt
flask==3.0.2
pytesseract==0.3.10
Pillow==10.2.0 